TAREA #10708 Sistema de calificaciones por países. Estructura preparada y funcionando.

+++ trunk/docroot/WEB-INF/src/com/liferay/lms/learningactivity/calificationtype/PercentCalificationType.java
@@ -33,7 +33,7 @@ public class PercentCalificationType extends BaseCalificationType {
 	
 	@Override
 	public String translate(double result) {
-		return Long.toString(result);
+		return Double.toString(result);
 	}
 	
 }
